Add currency to schemas
- Updated schemas for items
- Updated namespace for item transformers
- Added currency code to default item collection
