+++ app/Entity/Item/SimpleItem.php
@@ -62,7 +62,7 @@ class SimpleItem extends Item
 
     public function transformer(array $data_to_transform): Transformer
     {
-        return new \App\Models\Transformers\ItemType\SimpleItem($data_to_transform);
+        return new \App\Models\Transformers\Item\SimpleItem($data_to_transform);
     }
 
     public function update(array $patch, Model $instance): bool
