+++ app/Http/Controllers/ItemTypeView.php
@@ -10,7 +10,7 @@ use App\Response\Header\Headers;
 use App\Request\Parameter;
 use App\Request\Route;
 use App\Response\Pagination as UtilityPagination;
-use App\Models\Transformers\ItemType as ItemTypeTransformer;
+use App\Models\Transformers\Item as ItemTypeTransformer;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
