+++ public/api/schema/item-allocated-expense-include-categories.json
@@ -10,6 +10,11 @@
         "description": {
             "type": ["null", "string"]
         },
+        "currency_code": {
+            "type": "string",
+            "maxLength": 3,
+            "minLength": 3
+        },
         "total": {
             "type": "string"
         },
@@ -61,6 +66,7 @@
         "id",
         "name",
         "description",
+        "currency_code",
         "total",
         "percentage",
         "actualised_total",
