+++ public/api/schema/item-allocated-expense-include-category.json
@@ -10,6 +10,11 @@
         "description": {
             "type": ["null", "string"]
         },
+        "currency_code": {
+            "type": "string",
+            "maxLength": 3,
+            "minLength": 3
+        },
         "total": {
             "type": "string"
         },
@@ -44,6 +49,7 @@
         "id",
         "name",
         "description",
+        "currency_code",
         "total",
         "percentage",
         "actualised_total",
