+++ public/api/schema/item-simple-expense-include-category.json
@@ -10,6 +10,11 @@
         "description": {
             "type": "null"
         },
+        "currency_code": {
+            "type": "string",
+            "maxLength": 3,
+            "minLength": 3
+        },
         "total": {
             "type": "string"
         },
@@ -38,6 +43,7 @@
         "id",
         "name",
         "description",
+        "currency_code",
         "total",
         "created",
         "category",
